Added Vector.h file from my ray marching project and adjusted it.

+++ main.cpp
@@ -33,7 +33,7 @@ void Run()
 
 		Render();
 
-		SDL_Delay(timestep);
+		SDL_Delay((Uint32)timestep);
 	}
 }
